fix(environment-widget): extract service logic from DeploymentsService
Extract EnvironmentWidget-specific function and interfaces out of DeploymentsService and into new
ApplicationOverviewService, specfic to the EnvironmentWidgetModule, for better separation and
encapsulation between the Deployments and dashboard widgets modules.

fixes https://github.com/openshiftio/openshift.io/issues/2961
