+++ src/app/dashboard-widgets/environment-widget/environment-widget.component.ts
@@ -8,15 +8,15 @@ import { Context, Contexts, Spaces } from 'ngx-fabric8-wit';
 import { Space } from '../../../app/space/create/deployments/services/deployment-api.service';
 import {
   ApplicationAttributesOverview,
-  DeploymentsService
-} from '../../../app/space/create/deployments/services/deployments.service';
+  ApplicationOverviewService
+} from './application-overview.service';
 
 @Component({
   encapsulation: ViewEncapsulation.None,
   selector: 'fabric8-environment-widget',
   templateUrl: './environment-widget.component.html',
   styleUrls: ['./environment-widget.component.less'],
-  providers: [DeploymentsService]
+  providers: [ApplicationOverviewService]
 })
 export class EnvironmentWidgetComponent implements OnInit {
 
@@ -26,13 +26,13 @@ export class EnvironmentWidgetComponent implements OnInit {
 
   constructor(private context: Contexts,
               private spaces: Spaces,
-              private deploymentsService: DeploymentsService) {
+              private applicationOverviewService: ApplicationOverviewService) {
     this.spaceId = this.spaces.current.first().map(space => space.id);
   }
 
   ngOnInit() {
     this.spaceId.subscribe((spaceId: string) => {
-      this.appInfos = this.deploymentsService.getAppsAndEnvironments(spaceId);
+      this.appInfos = this.applicationOverviewService.getAppsAndEnvironments(spaceId);
     });
 
     this.contextPath = this.context.current.map(context => context.path);
